# openapi.yaml
openapi: 3.0.3
info:
  title: SportsDeck API - Premier League
  description: |
    Backend API for SportsDeck, a Premier League discussion platform.
    Features include authentication, match data (via football-data.org), discussion threads,
    polls, moderation with AI-powered content analysis, social features, and AI enhancements.
  version: 1.0.0
  contact:
    name: SportsDeck
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login or /api/auth/signup
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        avatar:
          type: string
          nullable: true
        role:
          type: string
          enum: [USER, ADMIN]
        isBanned:
          type: boolean
        favoriteTeamId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
paths:
  /api/auth/signup:
    post:
      tags: [Auth]
      summary: Register a new user
      description: Create a new account with email, password, and username. Returns JWT tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, username]
              properties:
                email:
                  type: string
                  example: testuser@example.com
                password:
                  type: string
                  example: password123
                username:
                  type: string
                  example: testuser
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
              example:
                user:
                  id: 1
                  email: testuser@example.com
                  username: testuser
                  role: USER
                  createdAt: "2026-03-01T00:00:00.000Z"
                accessToken: "eyJhbGciOiJIUzI1NiIs..."
                refreshToken: "eyJhbGciOiJIUzI1NiIs..."
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email or username already in use
...